Add prompt and info logs to setup script

+++ setup.sh
@@ -1,36 +1,57 @@
-rm -f $HOME/.aliases
-ln -s `pwd`/dot_aliases $HOME/.aliases
+read -p "This command will delete existing configs from home directory. Are you sure? (y/n)" -n 1 -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    echo "Creating .aliases"
+    rm -f $HOME/.aliases
+    ln -s `pwd`/dot_aliases $HOME/.aliases
 
-rm -f $HOME/.gitconfig
-ln -s `pwd`/dot_gitconfig $HOME/.gitconfig
+    echo "Creating .gitconfig"
+    rm -f $HOME/.gitconfig
+    ln -s `pwd`/dot_gitconfig $HOME/.gitconfig
 
-rm -f $HOME/.gitconfig-work
-ln -s `pwd`/dot_gitconfig_work $HOME/.gitconfig-work
+    echo "Creating .gitconfig-work"
+    rm -f $HOME/.gitconfig-work
+    ln -s `pwd`/dot_gitconfig_work $HOME/.gitconfig-work
 
-rm -f $HOME/.rubocop.yml
-ln -s `pwd`/dot_rubocop.yml $HOME/.rubocop.yml
+    echo "Creating .rubocop.yml"
+    rm -f $HOME/.rubocop.yml
+    ln -s `pwd`/dot_rubocop.yml $HOME/.rubocop.yml
 
-rm -f $HOME/.utilities
-ln -s `pwd`/dot_utilities $HOME/.utilities
+    echo "Creating .utilities"
+    rm -f $HOME/.utilities
+    ln -s `pwd`/dot_utilities $HOME/.utilities
 
-rm -f $HOME/.zshenv
-ln -s `pwd`/dot_zshenv $HOME/.zshenv
+    echo "Creating .zshenv"
+    rm -f $HOME/.zshenv
+    ln -s `pwd`/dot_zshenv $HOME/.zshenv
 
-rm -f $HOME/.zim_preload
-ln -s `pwd`/dot_zim_preload $HOME/.zim_preload
+    echo "Creating .zim_preload"
+    rm -f $HOME/.zim_preload
+    ln -s `pwd`/dot_zim_preload $HOME/.zim_preload
 
-rm -f $HOME/.zlogin
-ln -s `pwd`/dot_zlogin $HOME/.zlogin
+    echo "Creating .zlogin"
+    rm -f $HOME/.zlogin
+    ln -s `pwd`/dot_zlogin $HOME/.zlogin
 
-rm -f $HOME/.zimrc
-ln -s `pwd`/dot_zimrc $HOME/.zimrc
+    echo "Creating .zimrc"
+    rm -f $HOME/.zimrc
+    ln -s `pwd`/dot_zimrc $HOME/.zimrc
 
-rm -f $HOME/.zshrc
-ln -s `pwd`/dot_zshrc $HOME/.zshrc
+    echo "Creating .zshrc"
+    rm -f $HOME/.zshrc
+    ln -s `pwd`/dot_zshrc $HOME/.zshrc
 
-rm -f $HOME/.sqliterc
-ln -s `pwd`/dot_sqliterc $HOME/.sqliterc
+    echo "Creating .sqliterc"
+    rm -f $HOME/.sqliterc
+    ln -s `pwd`/dot_sqliterc $HOME/.sqliterc
 
-mkdir -p $HOME/.ghc
-rm -f $HOME/.ghc/ghci.conf
-ln -s `pwd`/ghci.conf $HOME/.ghc/ghci.conf
+    echo "Creating .vimrc"
+    rm -f $HOME/.vimrc
+    ln -s `pwd`/dot_vimrc $HOME/.vimrc
+
+    echo "Creating .ghc/ghci.conf"
+    mkdir -p $HOME/.ghc
+    rm -f $HOME/.ghc/ghci.conf
+    ln -s `pwd`/ghci.conf $HOME/.ghc/ghci.conf
+fi
